+++ NEWS.md
@@ -10,6 +10,9 @@ listing of changes. See the git commit log for additional details.
 
 # 2021-06-xx
 
+# Highlights:
+- Fixed building outside Nix sandbox (e.g. in Docker) by introducing `signing.buildTimeKeyStorePath` option.
+
 ## Backwards incompatible changes
 - Renamed `keyStorePath` to `signing.keyStorePath`.